Copy emitter.
