+++ crates/cli/Cargo.toml
@@ -30,7 +30,6 @@ moon_pipeline_provider = { path = "../core/pipeline-provider" }
 moon_platform = { path = "../core/platform" }
 moon_project = { path = "../core/project" }
 moon_project_graph = { path = "../core/project-graph" }
-moon_runner = { path = "../core/runner" }
 moon_system_platform = { path = "../system/platform" }
 moon_task = { path = "../core/task" }
 moon_terminal = { path = "../core/terminal" }
